+++ src/generators/handler.ts
@@ -1,9 +1,9 @@
-import { WorkspaceConfiguration } from "vscode";
+import type { WorkspaceConfiguration } from "vscode";
 
 export const generateHandler = (config: WorkspaceConfiguration) => {
   const customHandle = config.get("handle");
   if (customHandle) {
-    return customHandle + "\n";
+    return `${customHandle}\n`;
   }
-  return [`export const handle = () => ({`, `  // your handler here`, `});`].join("\n");
+  return ["export const handle = () => ({", "  // your handler here", "});"].join("\n");
 };